requirements.txt
cached-property==1.3.1
Django==2.0.2
django-currentuser==0.2.2
django-url-filter==0.3.4
djangorestframework==3.7.7
enum-compat==0.0.2
pytz==2017.3
six==1.11.0
django-url-filter
